@@ -261,7 +261,7 @@ func (s *decimalTestSuite) TestArithmetic() {
 	}
 
 	for tcIndex, tc := range tests {
-		tc := tc
+
 		resAdd := tc.d1.Add(tc.d2)
 		resSub := tc.d1.Sub(tc.d2)
 		resMul := tc.d1.Mul(tc.d2)
@@ -727,7 +727,6 @@ func TestFormatDec(t *testing.T) {
 	require.NoError(t, err)
 
 	for _, tc := range testcases {
-		tc := tc
 		t.Run(tc[0], func(t *testing.T) {
 			out, err := math.FormatDec(tc[0])
 			require.NoError(t, err)
