@@ -39,9 +39,18 @@ func (a *AppBuilder) Build(db corestore.KVStoreWithBatch, traceStore io.Writer, 
 		baseAppOptions = append(baseAppOptions, option)
 	}
 
+	// set routers first in case they get modified by other options
+	baseAppOptions = append(
+		[]func(*baseapp.BaseApp){
+			func(bApp *baseapp.BaseApp) {
+				bApp.SetMsgServiceRouter(a.app.msgServiceRouter)
+				bApp.SetGRPCQueryRouter(a.app.grpcQueryRouter)
+			},
+		},
+		baseAppOptions...,
+	)
+
 	bApp := baseapp.NewBaseApp(a.app.config.AppName, a.app.logger, db, nil, baseAppOptions...)
-	bApp.SetMsgServiceRouter(a.app.msgServiceRouter)
-	bApp.SetGRPCQueryRouter(a.app.grpcQueryRouter)
 	bApp.SetCommitMultiStoreTracer(traceStore)
 	bApp.SetVersion(version.Version)
 	bApp.SetInterfaceRegistry(a.app.interfaceRegistry)
