@@ -139,7 +139,7 @@ func ListenerMux(listeners ...Listener) Listener {
 
 	mux.onBatch = func(batch PacketBatch) error {
 		for _, listener := range listeners {
-			err := batch.apply(&listener)
+			err := batch.apply(&listener) //nolint:gosec // aliasing is safe here
 			if err != nil {
 				return err
 			}
