@@ -1,6 +1,2 @@
-/*
-The mempool package defines a few mempool services which can be used in conjunction with your consensus implementation
-
-*/
-
+// Package mempool defines a few mempool services which can be used in conjunction with your consensus implementation.
 package mempool
