@@ -9,14 +9,15 @@ import (
 )
 
 // ServerOptions defines the options for the CometBFT server.
+// When an option takes a map[string]any, it can access the app.tom's cometbft section and the config.toml config.
 type ServerOptions[T transaction.Tx] struct {
-	Mempool                    mempool.Mempool[T]
 	PrepareProposalHandler     handlers.PrepareHandler[T]
 	ProcessProposalHandler     handlers.ProcessHandler[T]
 	VerifyVoteExtensionHandler handlers.VerifyVoteExtensionhandler
 	ExtendVoteHandler          handlers.ExtendVoteHandler
 
-	SnapshotOptions snapshots.SnapshotOptions
+	Mempool         func(cfg map[string]any) mempool.Mempool[T]
+	SnapshotOptions func(cfg map[string]any) snapshots.SnapshotOptions
 
 	AddrPeerFilter types.PeerFilter // filter peers by address and port
 	IdPeerFilter   types.PeerFilter // filter peers by node ID
@@ -26,12 +27,12 @@ type ServerOptions[T transaction.Tx] struct {
 // It defaults to a NoOpMempool and NoOp handlers.
 func DefaultServerOptions[T transaction.Tx]() ServerOptions[T] {
 	return ServerOptions[T]{
-		Mempool:                    mempool.NoOpMempool[T]{},
 		PrepareProposalHandler:     handlers.NoOpPrepareProposal[T](),
 		ProcessProposalHandler:     handlers.NoOpProcessProposal[T](),
 		VerifyVoteExtensionHandler: handlers.NoOpVerifyVoteExtensionHandler(),
 		ExtendVoteHandler:          handlers.NoOpExtendVote(),
-		SnapshotOptions:            snapshots.NewSnapshotOptions(0, 0),
+		Mempool:                    func(cfg map[string]any) mempool.Mempool[T] { return mempool.NoOpMempool[T]{} },
+		SnapshotOptions:            func(cfg map[string]any) snapshots.SnapshotOptions { return snapshots.NewSnapshotOptions(0, 0) },
 		AddrPeerFilter:             nil,
 		IdPeerFilter:               nil,
 	}
