@@ -38,14 +38,14 @@ func AddCommands[T transaction.Tx](
 	rootCmd *cobra.Command,
 	newApp AppCreator[T],
 	logger log.Logger,
-	serverCfg ServerConfig,
+	globalServerCfg ServerConfig,
 	components ...ServerComponent[T],
 ) error {
 	if len(components) == 0 {
 		return errors.New("no components provided")
 	}
 
-	server := NewServer(logger, serverCfg, components...)
+	server := NewServer(logger, globalServerCfg, components...)
 	originalPersistentPreRunE := rootCmd.PersistentPreRunE
 	rootCmd.PersistentPreRunE = func(cmd *cobra.Command, args []string) error {
 		// set the default command outputs
