@@ -266,10 +266,8 @@ var (
 				Config: appconfig.WrapAny(&govmodulev1.Module{}),
 			},
 			{
-				Name: consensustypes.ModuleName,
-				Config: appconfig.WrapAny(&consensusmodulev1.Module{
-					Authority: "consensus", // TODO remove.
-				}),
+				Name:   consensustypes.ModuleName,
+				Config: appconfig.WrapAny(&consensusmodulev1.Module{}),
 			},
 			{
 				Name:   accounts.ModuleName,
