@@ -169,7 +169,6 @@ func (s *TestSuite) TestCreateGroup() {
 
 	var seq uint32 = 1
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			blockTime := sdk.UnwrapSDKContext(s.ctx).HeaderInfo().Time
 			res, err := s.groupKeeper.CreateGroup(s.ctx, spec.req)
@@ -520,7 +519,6 @@ func (s *TestSuite) TestUpdateGroupMembers() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			sdkCtx, _ := s.sdkCtx.CacheContext()
 			_, err := s.groupKeeper.UpdateGroupMembers(sdkCtx, spec.req)
@@ -657,7 +655,6 @@ func (s *TestSuite) TestUpdateGroupAdmin() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			_, err := s.groupKeeper.UpdateGroupAdmin(s.ctx, spec.req)
 			if spec.expErr {
@@ -727,7 +724,6 @@ func (s *TestSuite) TestUpdateGroupMetadata() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			sdkCtx, _ := s.sdkCtx.CacheContext()
 			_, err := s.groupKeeper.UpdateGroupMetadata(sdkCtx, spec.req)
@@ -897,7 +893,6 @@ func (s *TestSuite) TestCreateGroupWithPolicy() {
 	}
 
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			s.setNextAccount()
 			err := spec.req.SetDecisionPolicy(spec.policy)
@@ -1084,7 +1079,6 @@ func (s *TestSuite) TestCreateGroupPolicy() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			err := spec.req.SetDecisionPolicy(spec.policy)
 			s.Require().NoError(err)
@@ -1218,7 +1212,7 @@ func (s *TestSuite) TestUpdateGroupPolicyAdmin() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
+
 		err := spec.expGroupPolicy.SetDecisionPolicy(policy)
 		s.Require().NoError(err)
 
@@ -1369,7 +1363,7 @@ func (s *TestSuite) TestUpdateGroupPolicyDecisionPolicy() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
+
 		policyAddr := groupPolicyAddr
 		err := spec.expGroupPolicy.SetDecisionPolicy(spec.policy)
 		s.Require().NoError(err)
@@ -1470,7 +1464,7 @@ func (s *TestSuite) TestUpdateGroupPolicyMetadata() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
+
 		err := spec.expGroupPolicy.SetDecisionPolicy(policy)
 		s.Require().NoError(err)
 
@@ -1877,7 +1871,6 @@ func (s *TestSuite) TestSubmitProposal() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			err := spec.req.SetMsgs(spec.msgs)
 			s.Require().NoError(err)
@@ -2002,7 +1995,7 @@ func (s *TestSuite) TestWithdrawProposal() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
+
 		s.Run(msg, func() {
 			pID := spec.preRun(s.sdkCtx)
 
@@ -2385,7 +2378,6 @@ func (s *TestSuite) TestVote() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			sdkCtx := s.sdkCtx
 			if !spec.srcCtx.IsZero() {
@@ -2738,7 +2730,6 @@ func (s *TestSuite) TestExecProposal() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			sdkCtx, _ := s.sdkCtx.CacheContext()
 			proposalID := spec.setupProposal(sdkCtx)
@@ -2782,7 +2773,6 @@ func (s *TestSuite) TestExecProposal() {
 			}
 			spec.postRun(sdkCtx)
 		})
-
 	}
 }
 
@@ -2935,7 +2925,6 @@ func (s *TestSuite) TestExecPrunedProposalsAndVotes() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			sdkCtx, _ := s.sdkCtx.CacheContext()
 			proposalID := spec.setupProposal(sdkCtx)
@@ -3343,7 +3332,6 @@ func (s *TestSuite) TestExecProposalsWhenMemberLeavesOrIsUpdated() {
 		},
 	}
 	for msg, spec := range specs {
-		spec := spec
 		s.Run(msg, func() {
 			sdkCtx, _ := s.sdkCtx.CacheContext()
 
