@@ -32,6 +32,10 @@ type EncoderOptions struct {
 	// EnumAsString when set will encode enums as strings instead of integers.
 	// Caution: Enabling this option produce different sign bytes.
 	EnumAsString bool
+	// AminoNameAsTypeURL when set will use the amino name as the type URL in the JSON output.
+	// It is useful when using the Amino JSON encoder for non Amino purposes,
+	// such as JSON RPC.
+	AminoNameAsTypeURL bool
 	// TypeResolver is used to resolve protobuf message types by TypeURL when marshaling any packed messages.
 	TypeResolver signing.TypeResolver
 	// FileResolver is used to resolve protobuf file descriptors TypeURL when TypeResolver fails.
@@ -50,6 +54,7 @@ type Encoder struct {
 	doNotSortFields           bool
 	indent                    string
 	enumsAsString             bool
+	aminoNameAsTypeURL        bool
 }
 
 // NewEncoder returns a new Encoder capable of serializing protobuf messages to JSON using the Amino JSON encoding
@@ -80,11 +85,12 @@ func NewEncoder(options EncoderOptions) Encoder {
 			"google.protobuf.Duration":  marshalDuration,
 			"google.protobuf.Any":       marshalAny,
 		},
-		fileResolver:    options.FileResolver,
-		typeResolver:    options.TypeResolver,
-		doNotSortFields: options.DoNotSortFields,
-		indent:          options.Indent,
-		enumsAsString:   options.EnumAsString,
+		fileResolver:       options.FileResolver,
+		typeResolver:       options.TypeResolver,
+		doNotSortFields:    options.DoNotSortFields,
+		indent:             options.Indent,
+		enumsAsString:      options.EnumAsString,
+		aminoNameAsTypeURL: options.AminoNameAsTypeURL,
 	}
 	return enc
 }
@@ -187,9 +193,17 @@ func (enc Encoder) beginMarshal(msg protoreflect.Message, writer io.Writer, isAn
 	)
 
 	if isAny {
-		name, named = getMessageAminoNameAny(msg), true
+		if enc.aminoNameAsTypeURL {
+			name, named = getMessageTypeURL(msg), true
+		} else {
+			name, named = getMessageAminoNameAny(msg), true
+		}
 	} else {
 		name, named = getMessageAminoName(msg)
+		if enc.aminoNameAsTypeURL {
+			// do not override named
+			name = getMessageTypeURL(msg)
+		}
 	}
 
 	if named {
